% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_defs.R
\docType{class}
\name{Atom-class}
\alias{Atom-class}
\title{An S4 class to represent atoms.}
\description{
An Atom has slots for the output from the UMLS API, as specified by \href{https://documentation.uts.nlm.nih.gov/rest/atoms/#sample-output}{the NLM.}
}

