% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept.R
\name{codes}
\alias{codes}
\title{Get the codes for atoms from a list of vocabularies.}
\usage{
codes(concept, vocab_list)
}
\arguments{
\item{vocab_list}{}
}
\value{
A list of `(code, source)` pairs.
}
\description{
Get the codes for atoms from a list of vocabularies.
}

