% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_defs.R
\docType{class}
\name{ConceptRelation-class}
\alias{ConceptRelation-class}
\title{An S4 class to represent concept relations.}
\description{
These are a subclass of \code{Relation.}
}
\section{Slots}{

\describe{
\item{\code{rui}}{The relation identifier.}

\item{\code{headui}}{The identifier of the head of the relation (target).}

\item{\code{tailui}}{The identifier of the tail of the relation (source).}

\item{\code{suppressible}}{Whether or not this relation is suppressible.}

\item{\code{sourceui}}{The source identifier.}

\item{\code{relatedConceptURL}}{The API URL of the related concept.}
}}

