% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_defs.R
\docType{class}
\name{Definition-class}
\alias{Definition-class}
\title{An S4 class to represent definitions.}
\description{
An S4 class to represent definitions.
}
\section{Slots}{

\describe{
\item{\code{rootSource}}{The source of the definition.}

\item{\code{sourceOriginated}}{Whether or not the definition originated from the source.}

\item{\code{value}}{The definition.}
}}

