% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{definitions}
\alias{definitions}
\title{#' Get UMLS concept definitions.}
\usage{
definitions(CUI, sabs)
}
\arguments{
\item{CUI}{The input concept, CUI, or list of such things.}

\item{sab}{List of source vocabularies, comma separated.}
}
\value{
A \linkS4class{Definition} object or a list of \linkS4class{Definition} objects.
}
\description{
Retrieves the definitions of a given \linkS4class{Concept}, as per \href{https://documentation.uts.nlm.nih.gov/rest/definitions/}{the NLM}. This is vectorized over the input.
}

