% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{filter_source_level}
\alias{filter_source_level}
\title{Filter results to include source vocabularies with a particular restriction level.}
\usage{
filter_source_level(results, level, us = TRUE)
}
\arguments{
\item{results}{Search results. Should be a list of things with a \code{rootSource} attribute.}

\item{level}{The maximum restriction.}

\item{us}{Whether or not you are looking at US or non-US restriction levels.}
}
\value{
The results, filtered to include only those in source vocabularies with an appropriate restriction level.
}
\description{
See also \code{\link{source_level}}.
}
\examples{
# Get diabetic retinopathy results that are from source vocabularies with restriction level at most 1.
search <- search_UMLS("diabetic retinopathy")
filteredSearch <- search \%>\% filter_source_level(1)
}

