% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{get_all_children}
\alias{get_all_children}
\title{Get all of the children of a given source item.}
\usage{
get_all_children(source, id)
}
\arguments{
\item{source}{The source vocabulary.}

\item{id}{The internal source ID.}
}
\value{
A \code{data.tree} which contains the tree of children.
}
\description{
Get all of the children of a given source item.
}
\examples{
# Get all the CPT codes that are children of the CUI \\code{C0519273}.
conceptAtoms <- get_concept_atoms("C0519273")
cptCodes <- conceptAtoms \%>\% list.filter(rootSource == 'CPT') \%>\% list.select(code) \%>\% list.apply(function(x) rev(str_split(x, "/")[[1]])[[1]])
cptKids <- cptCodes \%>\% list.apply(function(x) get_all_children("CPT", x))
cptLeaves <- cptKids \%>\% list.apply(function(x) x$Get('id', filterFun = isLeaf)) \%>\% list.apply(function(x) data.frame(proc_code = x, desc = names(x))) \%>\% bind_rows()
}

