% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{get_all_parents}
\alias{get_all_parents}
\title{Get all of the parents of a given source item.}
\usage{
get_all_parents(source, id)
}
\arguments{
\item{source}{The source vocabulary.}

\item{id}{The internal source ID.}
}
\value{
A \code{data.tree} which contains the tree of parents.
}
\description{
Get all of the parents of a given source item.
}

