% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsets.R
\name{get_all_subset_info}
\alias{get_all_subset_info}
\title{Get information on all UMLS subsets.}
\usage{
get_all_subset_info()
}
\value{
A list of information about all subsets. These are of UMLS class \code{Subset}.
}
\description{
Get information on all UMLS subsets.
}

