% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{get_atom_family}
\alias{get_atom_ancestors}
\alias{get_atom_children}
\alias{get_atom_descendants}
\alias{get_atom_family}
\alias{get_atom_parents}
\title{Get familial relationships for a given Atom.}
\usage{
get_atom_family(AUI, type)

get_atom_parents(AUI)

get_atom_children(AUI)

get_atom_ancestors(AUI)

get_atom_descendants(AUI)
}
\arguments{
\item{AUI}{The AUI of interest.}

\item{type}{The type of familial relationship. This can be one of \code{ancestors, descendants, parents, children}. See the possible endpoints from the \href{https://documentation.uts.nlm.nih.gov/rest/atoms/}{UMLS}.}
}
\value{
A list of results. These are of UMLS class \code{Atom}.
}
\description{
Get familial relationships for a given Atom.
}
\examples{
# Get parents of atom A8345234
parents <- get_atom_parents("A8345234")
}

