% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{get_atom_info}
\alias{get_atom_info}
\title{Get information about a given atom.}
\usage{
get_atom_info(AUI)
}
\arguments{
\item{AUI}{The AUI of interest.}
}
\value{
Information about the atom. This is of UMLS class \code{Atom}.
}
\description{
Get information about a given atom.
}
\examples{
# Get information about atom A8345234
info <- get_atom_info("A8345234")
}

