% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept.R
\docType{methods}
\name{get_concept,character-method}
\alias{get_concept,character-method}
\title{Get Concept}
\usage{
\S4method{get_concept}{character}(x, info_ret = "none")
}
\arguments{
\item{x}{}
}
\description{
Get Concept
}

