% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_concept}
\alias{get_concept}
\title{Get UMLS Concept.}
\usage{
get_concept(CUI, info_ret = "none")
}
\arguments{
\item{CUI}{A CUI or list of CUIs.}
}
\value{
A \code{\linkS4class{Concept}} or list of  \code{\linkS4class{Concept}}s.
}
\description{
This retrieves a UMLS Concept, as identified by a CUI.
}

