% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept.R
\name{get_concept_atoms}
\alias{get_concept_atoms}
\alias{get_concept_atoms_page}
\title{Get UMLS concept atoms.}
\usage{
get_concept_atoms(CUI, sabs = NULL, ttys = NULL, language = NULL,
  includeObsolete = FALSE, includeSuppressible = FALSE)

get_concept_atoms_page(CUI, sabs = NULL, ttys = NULL, language = NULL,
  includeObsolete = FALSE, includeSuppressible = FALSE, pageNumber = 1,
  pageSize = 25)
}
\arguments{
\item{CUI}{CUI of interest.}

\item{sabs}{Source vocabularies, comma separated.}
}
\description{
Get UMLS concept atoms.
}

