% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept.R
\name{get_concept_defs}
\alias{get_concept_defs}
\title{Get UMLS concept definitions.}
\usage{
get_concept_defs(CUI, sabs = NULL)
}
\arguments{
\item{CUI}{CUI of interest.}

\item{sabs}{Source vocabularies, comma separated.}
}
\value{
A list of \linkS4class{Definition} objects.
}
\description{
Retrieves the definitions of a given \linkS4class{Concept}, as per \href{https://documentation.uts.nlm.nih.gov/rest/definitions/}{the NLM}.
}

