% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{get_concept_info}
\alias{get_concept_info}
\alias{get_concept_info_page}
\title{Get UMLS concept info}
\usage{
get_concept_info(CUI)

get_concept_info_page(CUI, pageNumber = 1, pageSize = 25)
}
\arguments{
\item{CUI}{CUI of interest.}

\item{sabs}{Source vocabularies, comma separated.}
}
\description{
Get UMLS concept info
}

