% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept.R
\name{get_concept_rels}
\alias{get_concept_rels}
\title{Get the Metathesaurus Concept-Concept relationships for a given CUI.}
\usage{
get_concept_rels(CUI)
}
\arguments{
\item{CUI}{}
}
\value{
A list of Concept-Concept relationships. These are of UMLS class \code{ConceptRelation}.
}
\description{
Get the Metathesaurus Concept-Concept relationships for a given CUI.
}
\examples{
# Get relationships for concept C0011884
relations <- get_concept_rels("C0011884")
}

