% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept.R
\name{get_internal_codes}
\alias{get_internal_codes}
\title{Get the internal codes for atoms from a list of vocabularies.}
\usage{
get_internal_codes(concept)
}
\arguments{
\item{concept}{}
}
\value{
A list of `(code, source)` pairs.
}
\description{
The source vocabularies in the UMLS have internal codes for the atoms. This will retrieve the \code{(code, source)} pairs that can be used to obtain internal information about the atoms.
}
\examples{
# Get the atoms about concept C0011884
concept <- get_concept_atoms("C0011884")
codePairs <- get_internal_codes(concept)
}

