% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsets.R
\name{get_mem_info}
\alias{get_mem_info}
\alias{get_subset_mems_page}
\title{Get information on a member of a given subset.}
\usage{
get_mem_info(subset, source, member, language = NULL)

get_subset_mems_page(subset, source, member, language = NULL,
  pageNumber = 1, pageSize = 25)
}
\arguments{
\item{language}{}
}
\description{
Get information on a member of a given subset.
}

