% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept.R
\name{get_pref_atom}
\alias{get_pref_atom}
\title{Get the preferred atom for a concept.}
\usage{
get_pref_atom(CUI)
}
\arguments{
\item{CUI}{CUI of interest.}
}
\value{
The preferred atom of the CUI. This is of UMLS class \code{Atom}.
}
\description{
Get the preferred atom for a concept.
}
\examples{
# Get the preferred atom of concept C0011884
prefAtom <- get_pref_atom("C0011884")
}

