% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_service_ticket}
\alias{get_service_ticket}
\alias{set_service_ticket}
\title{Get a service ticket.}
\usage{
get_service_ticket(TGT)

set_service_ticket(value)
}
\arguments{
\item{TGT}{The TGT token for the current session.}
}
\value{
Returns nothing, adding the service token to the environment.
}
\description{
Get a service ticket.
}

