% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{get_source_ancestors}
\alias{get_source_ancestors}
\alias{get_source_ancestors_page}
\title{Get ancestors of source object.}
\usage{
get_source_ancestors(source, id)

get_source_ancestors_page(source, id, pageNumber = 1, pageSize = 25)
}
\arguments{
\item{id}{}
}
\description{
Get ancestors of source object.
}

