% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{get_source_attrs}
\alias{get_source_attrs}
\alias{get_source_attrs_page}
\title{Get source-asserted attributes.}
\usage{
get_source_attrs(source, id)

get_source_attrs_page(source, id, pageNumber = 1, pageSize = 25)
}
\arguments{
\item{id}{}
}
\description{
Get source-asserted attributes.
}

