% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{get_source_children}
\alias{get_source_children}
\alias{get_source_children_page}
\title{Get children of source object.}
\usage{
get_source_children(source, id)

get_source_children_page(source, id, pageNumber = 1, pageSize = 25)
}
\arguments{
\item{id}{}
}
\description{
Get children of source object.
}

