% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{get_source_descendants}
\alias{get_source_descendants}
\alias{get_source_descendants_page}
\title{Get descendants of source object.}
\usage{
get_source_descendants(source, id)

get_source_descendants_page(source, id, pageNumber = 1, pageSize = 25)
}
\arguments{
\item{id}{}
}
\description{
Get descendants of source object.
}

