% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{get_source_family}
\alias{get_source_ancestors}
\alias{get_source_children}
\alias{get_source_descendants}
\alias{get_source_family}
\alias{get_source_parents}
\title{Get source familial relationships.}
\usage{
get_source_family(source, id, type)

get_source_parents(source, id)

get_source_children(source, id)

get_source_ancestors(source, id)

get_source_descendants(source, id)
}
\arguments{
\item{source}{The source vocabulary.}

\item{id}{The internal source ID.}

\item{type}{The type of familial relationship to obtain. Can be one of \code{parents, children, ancestors, descendants}. See the possible \code{/source/id/} endpoints at the UMLS \href{https://documentation.uts.nlm.nih.gov/rest/home.html}{UMLS}.}
}
\value{
A list of results. These are of class \code{SourceAtomCluster} in the UMLS.
}
\description{
Get source familial relationships.
}

