% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{get_source_parents}
\alias{get_source_parents}
\alias{get_source_parents_page}
\title{Get parents of source object.}
\usage{
get_source_parents(source, id)

get_source_parents_page(source, id, pageNumber = 1, pageSize = 25)
}
\arguments{
\item{id}{}
}
\description{
Get parents of source object.
}

