% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsets.R
\name{get_subset_attr}
\alias{get_subset_attr}
\alias{get_subset_attr_page}
\title{Get the attributes of a UMLS subset.}
\usage{
get_subset_attr(source, subset, language = NULL)

get_subset_attr_page(source, subset, language = NULL, pageNumber = 1,
  pageSize = 25)
}
\arguments{
\item{language}{}
}
\description{
Get the attributes of a UMLS subset.
}

