% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsets.R
\name{get_all_subset_info_page}
\alias{get_all_subset_info_page}
\alias{get_subset_info}
\alias{get_subset_info_page}
\title{Get the info on a UMLS subset.}
\usage{
get_all_subset_info_page(pageNumber = 1, pageSize = 25)

get_subset_info(source, subset, language = NULL)

get_subset_info_page(source, subset, language = NULL, pageNumber = 1,
  pageSize = 25)
}
\arguments{
\item{source}{}
}
\description{
Get the info on a UMLS subset.
}

