% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_TGT}
\alias{get_TGT}
\alias{reset_TGT}
\alias{set_TGT}
\title{Get a TGT token from the UMLS API.}
\usage{
get_TGT(name = NULL, pass = NULL)

set_TGT(value)

reset_TGT()
}
\arguments{
\item{name}{The UMLS username.}

\item{pass}{The UMLS password.}
}
\value{
Returns nothing, adding the TGT token to the environment.
}
\description{
TGT tokens are used to generate service tickets that allow for requests to be ran against the UMLS API.
}

