% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{headui}
\alias{headui}
\title{Get the head UI for a relation.}
\usage{
headui(x)
}
\value{
The UI of the target of a relationship, or a list of such things.
}
\description{
Get the head UI for a relation.
}

