% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\docType{methods}
\name{relation_label,Relation-method}
\alias{relation_label,Relation-method}
\title{Get a relation label.}
\usage{
\S4method{relation_label}{Relation}(rel, type = "primary")
}
\arguments{
\item{type}{}
}
\description{
Get a relation label.
}

