% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{relation_label}
\alias{relation_label}
\title{Get labels for relationship.}
\usage{
relation_label(rel, type)
}
\arguments{
\item{rel}{The input relation.}

\item{type}{The type of relation.}
}
\value{
A label for the relation or a list of such things.
}
\description{
Retrieves the labels for a relationship \code{rel}. Vectorized over \code{rel}.
}

