% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{relations}
\alias{relations}
\title{Get the relations associated to an object.}
\usage{
relations(x)
}
\value{
A \linkS4class{Relation} object or list of such things.
}
\description{
Get the relations associated to an object.
}

