% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rUMLS.R
\docType{package}
\name{rUMLS}
\alias{rUMLS}
\alias{rUMLS-package}
\title{An R Interface To The NLM Unified Medical Language System}
\description{
This package provides an interface to the NLM's UMLS. You will need an UMLS license (with the corresponding username and password) to access it. For more information, please see \url{https://www.nlm.nih.gov/research/umls/}.
}

