% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_UMLS}
\alias{parse_search}
\alias{search_UMLS}
\alias{search_UMLS_page}
\title{Search the UMLS.}
\usage{
search_UMLS(search, inputType = "sourceUi", includeObsolete = FALSE,
  includeSuppressible = FALSE, sabs = NULL, searchType = "words")

search_UMLS_page(search, inputType = "sourceUi", includeObsolete = FALSE,
  includeSuppressible = FALSE, sabs = NULL, searchType = "words",
  pageNumber = 1, pageSize = 25)

parse_search(result)
}
\arguments{
\item{search}{Input to search.}

\item{sabs}{Source vocabularies, comma delimited.}

\item{searchType}{}
}
\description{
This searches the UMLS for concepts.
}

