% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{source_level}
\alias{source_level}
\title{Obtain source vocabularies with a given licensing restriction level.}
\usage{
source_level(level, us = TRUE)
}
\arguments{
\item{level}{The maximum restriction level.}

\item{us}{Whether or not you are looking at US or non-US restriction levels.}
}
\value{
A list of source vocabularies that have restriction levels less than the entered level.
}
\description{
Various UMLS source vocabularies have licensing restrictions, ranging from none (0) to level 4. See \href{https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/license_agreement_appendix.html}{UMLS} for details.
}
\examples{
# Get all restriction-less US source vocabularies
source_level(0)
}

