% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{source_vocab}
\alias{source_vocab}
\title{Get the source vocabulary for an object.}
\usage{
source_vocab(x)
}
\value{
The source vocabulary for an object, or a list of such things.
}
\description{
Get the source vocabulary for an object.
}

