% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{tailui}
\alias{tailui}
\title{Get the tail UI for a relation.}
\usage{
tailui(x)
}
\value{
The UI of the source of a relationship, or a list of such things.
}
\description{
Get the tail UI for a relation.
}

