% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{UMLS_creds}
\alias{UMLS_creds}
\title{Get or set UMLS credentials.}
\usage{
UMLS_creds(force = FALSE)
}
\description{
\code{UMLS_creds} sets or gets the UMLS credentials for the current session.
}
\details{
If the environment vars \code{UMLS_USER, UMLS_PASS} have been set in \code{.Renviron}, they will be loaded. Otherwise, you will be prompted to enter a username/password.
}

