% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_defs.R
\docType{class}
\name{Atom-class}
\alias{Atom-class}
\title{An S4 class to represent atoms.}
\description{
An Atom has slots for the output from the UMLS API, as specified by \href{https://documentation.uts.nlm.nih.gov/rest/atoms/#sample-output}{the NLM.}
}
\section{Slots}{

\describe{
\item{\code{aui}}{The atom identifier.}

\item{\code{suppressible}}{Whether or not this atom is suppressible.}

\item{\code{rootSource}}{The root source vocabulary of the atom.}

\item{\code{termType}}{The type of term that the atom represents.}

\item{\code{codeURL}}{The API URL for the atom's code.}

\item{\code{conceptURL}}{The API URL for the atom's concept.}

\item{\code{sourceConceptURL}}{The API URL for the atom's source concept.}

\item{\code{sourceDescriptor}}{The descriptor from the source.}

\item{\code{attributesURL}}{The API URL for the atom's source attributes.}

\item{\code{parentsURL}}{The API URL for the atom's parents.}

\item{\code{childrenURL}}{The API URL for the atom's parents.}

\item{\code{relations}}{The source-asserted relations for the atom.}

\item{\code{name}}{The source-asserted atom name.}

\item{\code{language}}{The language of the atom.}

\item{\code{definitionsURL}}{The API URL for the atom's definition.}
}}

