% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_defs.R
\docType{class}
\name{Concept-class}
\alias{Concept-class}
\title{An S4 class to represent concepts.}
\description{
A Concept has slots for the output from the UMLS API, as specified by  \href{https://documentation.uts.nlm.nih.gov/rest/concept/#sample-output}{the NLM.}
}

