% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept.R
\name{get_concept_atoms}
\alias{get_concept_atoms}
\alias{get_concept_atoms_page}
\title{Get UMLS concept atoms.}
\usage{
get_concept_atoms(CUI, sabs = NULL, ttys = NULL, language = NULL,
  includeObsolete = FALSE, includeSuppressible = FALSE)

get_concept_atoms_page(CUI, sabs = NULL, ttys = NULL, language = NULL,
  includeObsolete = FALSE, includeSuppressible = FALSE, pageNumber = 1,
  pageSize = 25)
}
\arguments{
\item{CUI}{CUI of concept to obtain atoms for.}

\item{sabs}{Source vocabularies, comma separated.}

\item{ttys}{Term types, any one of the \href{https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/abbreviations.html#TTYC}{valid term types.}}

\item{language}{The language of interest.}

\item{includeObsolete}{Whether or not to include obsolete atoms.}

\item{includeSuppressible}{Whether or not to include suppressible atoms.}
}
\value{
A list of \linkS4class{Atom} objects.
}
\description{
Retrieves the atoms that are associated to a given \linkS4class{Concept}.
}

