% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{get_concept_defs}
\alias{get_concept_defs}
\title{Get UMLS concept definitions.}
\usage{
get_concept_defs(CUI, sabs = NULL)
}
\arguments{
\item{CUI}{CUI of interest.}

\item{sabs}{Source vocabularies, comma separated.}
}
\description{
Get UMLS concept definitions.
}

