% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_getters.R
\name{get_concept_rels}
\alias{get_concept_rels}
\alias{get_concept_rels_page}
\title{Get UMLS concept relations.}
\usage{
get_concept_rels(CUI)

get_concept_rels_page(CUI, pageNumber = 1, pageSize = 25)
}
\arguments{
\item{CUI}{}
}
\description{
Get UMLS concept relations.
}

