% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{get_concept_rels}
\alias{get_concept_rels}
\title{Get UMLS concept relations.}
\usage{
get_concept_rels(CUI, pageNumber = 1, pageSize = 25)
}
\arguments{
\item{pageSize}{}
}
\description{
Get UMLS concept relations.
}

