% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsets.R
\name{get_mem_info}
\alias{get_mem_info}
\title{Get information on a member of a given subset.}
\usage{
get_mem_info(subset, source, member, language = NULL)
}
\arguments{
\item{subset}{The subset.}

\item{source}{The source.}

\item{member}{The member.}

\item{language}{The desired language.}
}
\value{
A list of attributes. These are of UMLS class \code{Attribute}.
}
\description{
Get information on a member of a given subset.
}

