% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{get_source_atoms}
\alias{get_source_atoms}
\title{Get UMLS source info for a concept.}
\usage{
get_source_atoms(source, id, sabs = NULL, ttys = NULL, language = NULL,
  includeObsolete = FALSE, includeSuppressible = FALSE, preferred = FALSE)
}
\arguments{
\item{source}{Source vocabulary.}

\item{id}{Source ID.}
}
\description{
Get UMLS source info for a concept.
}

