% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{get_source_info}
\alias{get_source_info}
\title{Get UMLS source info for a concept.}
\usage{
get_source_info(source, id)
}
\arguments{
\item{source}{Source vocabulary.}

\item{id}{Source ID.}
}
\description{
Get UMLS source info for a concept.
}

