% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsets.R
\name{get_subset_attr}
\alias{get_subset_attr}
\title{Get the attributes of a UMLS subset.}
\usage{
get_subset_attr(source, subset, language = NULL)
}
\arguments{
\item{language}{}
}
\value{
A list of attributes. These are of UMLS class \code{Attribute}.
}
\description{
Get the attributes of a UMLS subset.
}

