% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsets.R
\name{get_all_subset_info_page}
\alias{get_all_subset_info_page}
\alias{get_subset_info}
\title{Get the info on a UMLS subset.}
\usage{
get_all_subset_info_page(pageNumber = 1, pageSize = 25)

get_subset_info(sourceVocab, subset, language = NULL)
}
\arguments{
\item{sourceVocab}{}
}
\value{
A list of information. This is of UMLS class \code{Subset}.
}
\description{
Get the info on a UMLS subset.
}

