% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_UMLS}
\alias{search_UMLS}
\title{Search the UMLS.}
\usage{
search_UMLS(search, inputType = "sourceUi", includeObsolete = FALSE,
  includeSuppressible = FALSE, sabs = NULL, searchType = "words",
  pageNumber = 1, pageSize = 25)
}
\arguments{
\item{search}{Input to search.}

\item{sabs}{Source vocabularies, comma delimited.}

\item{pageNumber}{Page number for pagination.}

\item{pageSize}{Size of page.}
}
\description{
This searches the UMLS for concepts.
}

