% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_UMLS}
\alias{search_UMLS}
\title{Search the UMLS.}
\usage{
search_UMLS(search, inputType = "sourceUi", includeObsolete = FALSE,
  includeSuppressible = FALSE, sabs = NULL, searchType = "words",
  response_type = "list")
}
\arguments{
\item{search}{The search input.}

\item{inputType}{The type of input. Can be one of}

\item{includeObsolete}{Whether or not to include obsolete results.}

\item{includeSuppressible}{Whether or not to include suppressible results.}

\item{sabs}{Source vocabularies to search.}

\item{searchType}{Type of search.}
}
\value{
If \code{response_type == 'list'}, then a list of results. Otherwise, returns a list of concepts with the\code{response_type} info included.
}
\description{
Search the UMLS.
}

