% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_UMLS}
\alias{search_UMLS}
\title{Search the UMLS.}
\usage{
search_UMLS(search, inputType = "sourceUi", includeObsolete = FALSE,
  includeSuppressible = FALSE, sabs = NULL, searchType = "words")
}
\arguments{
\item{search}{The search input.}

\item{inputType}{The type of input. Can be one of \code{'atom', 'code','sourceConcept','sourceDescriptor','sourceUi'} to obtain a valid response.}

\item{includeObsolete}{Whether or not to include obsolete results.}

\item{includeSuppressible}{Whether or not to include suppressible results.}

\item{sabs}{Source vocabularies to search. A list of elements from the \href{http://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/source_vocabularies.html}{valid source vocabularies.}}

\item{searchType}{Type of search. Can be one of \code{'exact','words','leftTruncation', 'rightTruncation','approximate', 'normalizedString'}.}
}
\value{
A list of results.
}
\description{
This allows for searching of the UMLS. The API parameters correspond to those on the \href{https://documentation.uts.nlm.nih.gov/rest/search/#query-parameters}{NLM website.}
}

